begin unification of compartment and process by moving compartment into process file.  circular imports are causing problems between process and repository.deriver_library
