+++ vivarium/compartments/chemotaxis_flagella.py
@@ -8,7 +8,7 @@ import argparse
 
 from vivarium.library.dict_utils import deep_merge
 from vivarium.library.units import units
-from vivarium.core.experiment import Compartment
+from vivarium.core.process import Compartment
 from vivarium.core.composition import (
     simulate_compartment_in_experiment,
     plot_simulation_output,
