+++ vivarium/compartments/chemotaxis_master.py
@@ -2,7 +2,7 @@ from __future__ import absolute_import, division, print_function
 
 import os
 
-from vivarium.core.experiment import Compartment
+from vivarium.core.process import Compartment
 from vivarium.core.composition import (
     simulate_compartment_in_experiment,
     plot_simulation_output,
