+++ vivarium/compartments/static_lattice.py
@@ -2,7 +2,7 @@ from __future__ import absolute_import, division, print_function
 
 import os
 
-from vivarium.core.experiment import Compartment
+from vivarium.core.process import Compartment
 from vivarium.core.composition import (
     compartment_in_experiment,
     COMPARTMENT_OUT_DIR,
