+++ vivarium/compartments/transport_metabolism.py
@@ -5,7 +5,7 @@ import argparse
 
 from vivarium.library.dict_utils import get_value_from_path
 from vivarium.library.units import units
-from vivarium.core.experiment import Compartment
+from vivarium.core.process import Compartment
 from vivarium.core.composition import (
     simulate_compartment_in_experiment,
     plot_simulation_output,
